Successful Deployment and Separation of Cargo
This project is beginning in earnest with the first separation of my main long script in Rust for a crate approach using Cargo. One Crate handles main game functions such as logic/rules, players, and dice, the other handles special features and spice to combat which will be supplemented in future releases. The readme is also updated to reflect changes.